+++ js/const.js
@@ -1,0 +1,59 @@
+const DESCRIPTIONS = [
+  'Прекрасный пейзаж!',
+  'А тут Кекс сам не свой',
+  'Отправились в поход несмотря на дождь',
+  'Долгожданная покупка машины!',
+  'Интересно, а сколько еще мы буем стоять в пробке?',
+  'Лучший музей Москвы!',
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const NAMES = [
+  'Никита Кожемяка',
+  'Василий',
+  'NaGiBaToR2003',
+  'Алина бьюти блогер',
+  'Анвар',
+  'Мистер Кредо',
+  'Анатолий Анатолий',
+  'Виктор Ильич',
+  'Дизайн интерьера Ростов',
+  'Дальнобой',
+  'Максим',
+  'Кекс',
+  'Лариса',
+  'Хмтл академович',
+  'Ольга',
+  'Алексей',
+  'не Алексей',
+  'Симпл',
+  'Димпл',
+  'Тамара',
+  'Харитон',
+  'Елена',
+  'Критик',
+  'Макар',
+  'Инстасамец',
+  'Александр',
+  'Бариста',
+  'Игорь',
+  'Ульяна',
+  'Замыкающий',
+];
+
+// Числовые константы
+
+const DESCRIPTIONS_COUNT = 25; // Количество описаний к фотографиям
+const MIN_LIKES = 15; // Минимальное количество лайков на фото
+const MAX_LIKES = 200; // Максимальное количество лайков на фото
+const COMMENTS_COUNT = 30; // Максимальное количество комментариев на фото
+
+export {DESCRIPTIONS, MESSAGES, NAMES, DESCRIPTIONS_COUNT, MIN_LIKES, MAX_LIKES, COMMENTS_COUNT};
